[Frontend] Relationship Radar widget refetch (#14448)
