--- opencti-platform/opencti-front/src/private/components/common/stix_relationships/StixRelationshipsRadar.jsx
@@ -1,191 +1,0 @@
-import React, { useState } from 'react';
-import { graphql } from 'react-relay';
-import { QueryRenderer } from '../../../../relay/environment';
-import { useFormatter } from '../../../../components/i18n';
-import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
-import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
-import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
-import WidgetRadar from '../../../../components/dashboard/WidgetRadar';
-import Loader, { LoaderVariant } from '../../../../components/Loader';
-import useDashboardViz from '../../../../components/dashboard/useDashboardViz';
-import WidgetNoHostEntity from '../../../../components/dashboard/WidgetNoHostEntity';
-
-const stixRelationshipsRadarsDistributionQuery = graphql`
-  query StixRelationshipsRadarDistributionQuery(
-    $field: String!
-    $operation: StatsOperation!
-    $startDate: DateTime
-    $endDate: DateTime
-    $dateAttribute: String
-    $isTo: Boolean
-    $limit: Int
-    $fromOrToId: [String]
-    $elementWithTargetTypes: [String]
-    $fromId: [String]
-    $fromRole: String
-    $fromTypes: [String]
-    $toId: [String]
-    $toRole: String
-    $toTypes: [String]
-    $relationship_type: [String]
-    $confidences: [Int]
-    $search: String
-    $filters: FilterGroup
-    $dynamicFrom: FilterGroup
-    $dynamicTo: FilterGroup
-  ) {
-    stixRelationshipsDistribution(
-      field: $field
-      operation: $operation
-      startDate: $startDate
-      endDate: $endDate
-      dateAttribute: $dateAttribute
-      isTo: $isTo
-      limit: $limit
-      fromOrToId: $fromOrToId
-      elementWithTargetTypes: $elementWithTargetTypes
-      fromId: $fromId
-      fromRole: $fromRole
-      fromTypes: $fromTypes
-      toId: $toId
-      toRole: $toRole
-      toTypes: $toTypes
-      relationship_type: $relationship_type
-      confidences: $confidences
-      search: $search
-      filters: $filters
-      dynamicFrom: $dynamicFrom
-      dynamicTo: $dynamicTo
-    ) {
-      label
-      value
-      entity {
-        ... on BasicObject {
-          id
-          entity_type
-        }
-        ... on BasicRelationship {
-          id
-          entity_type
-        }
-        ... on StixObject {
-          representative {
-            main
-          }
-        }
-        ... on StixRelationship {
-          representative {
-            main
-          }
-        }
-        # internal objects
-        ... on Creator {
-          name
-        }
-        ... on Group {
-          name
-        }
-        # need colors when available
-        ... on Label {
-          color
-        }
-        ... on MarkingDefinition {
-          x_opencti_color
-        }
-        ... on Status {
-          template {
-            name
-            color
-          }
-        }
-      }
-    }
-  }
-`;
-
-const StixRelationshipsRadar = ({
-  title,
-  variant,
-  height,
-  field,
-  startDate,
-  endDate,
-  dataSelection,
-  parameters = {},
-  popover,
-  host,
-}) => {
-  const { t_i18n } = useFormatter();
-  const [chart, setChart] = useState();
-  const { resolvedDataSelection, isMissingHostEntity, isPreviewMode } = useDashboardViz({
-    perspective: 'relationships',
-    dataSelection,
-    host,
-  });
-
-  const renderContent = () => {
-    if (isMissingHostEntity) {
-      return <WidgetNoHostEntity host={host} />;
-    }
-    let selection = {};
-    let filtersAndOptions;
-    if (resolvedDataSelection) {
-      selection = resolvedDataSelection[0];
-      filtersAndOptions = buildFiltersAndOptionsForWidgets(selection.filters, { isKnowledgeRelationshipWidget: true });
-    }
-    const finalField = selection.attribute || field || 'entity_type';
-    const variables = {
-      field: finalField,
-      operation: 'count',
-      startDate,
-      endDate,
-      dateAttribute: selection.date_attribute ?? 'created_at',
-      limit: selection.number ?? 10,
-      filters: filtersAndOptions?.filters,
-      isTo: selection.isTo,
-      dynamicFrom: selection.dynamicFrom,
-      dynamicTo: selection.dynamicTo,
-    };
-    return (
-      <QueryRenderer
-        query={stixRelationshipsRadarsDistributionQuery}
-        variables={variables}
-        render={({ props }) => {
-          if (
-            props
-            && props.stixRelationshipsDistribution
-            && props.stixRelationshipsDistribution.length > 0
-          ) {
-            return (
-              <WidgetRadar
-                data={props.stixRelationshipsDistribution}
-                label={selection.label}
-                groupBy={finalField}
-                onMounted={setChart}
-              />
-            );
-          }
-          if (props) {
-            return <WidgetNoData />;
-          }
-          return <Loader variant={LoaderVariant.inElement} />;
-        }}
-      />
-    );
-  };
-  return (
-    <WidgetContainer
-      padding="small"
-      height={height}
-      title={parameters.title ?? title ?? t_i18n('Relationships distribution')}
-      variant={variant}
-      chart={chart}
-      action={popover}
-      showPreviewTag={isPreviewMode}
-    >
-      {renderContent()}
-    </WidgetContainer>
-  );
-};
-
-export default StixRelationshipsRadar;
